move SequenceCounter to wire/rpl.rs
Use SequenceCounter in a parsed RPL packet.
